+++ src/photon_mosaic/core/selectplanes.py
